Shift ComposeService construction into DockerComposeServiceResource
